Users - Wallet - Crypto entities

+++ src/main/java/com/sensei/service/UserDbService.java
@@ -1,12 +1,15 @@
 package com.sensei.service;
 
 import com.sensei.entity.User;
+import com.sensei.entity.WalletCrypto;
+import com.sensei.exception.NotEmptyWalletException;
 import com.sensei.exception.UserNotFoundException;
 import com.sensei.repository.UserDao;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -31,7 +34,14 @@ public class UserDbService {
         return userDao.save(user);
     }
 
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id) throws UserNotFoundException, NotEmptyWalletException {
+        User user = userDao.findById(id).orElseThrow(UserNotFoundException::new);
+        List<WalletCrypto> filteredList = user.getWallet().getCryptosList().stream()
+                        .filter(n  -> n.getQuantity().doubleValue() != 0.00)
+                                .collect(Collectors.toList());
+        if(!filteredList.isEmpty()) {
+            throw new NotEmptyWalletException();
+        }
         userDao.deleteById(id);
     }
 }
